+++ CHANGELOG.md
@@ -33,6 +33,7 @@ All notable changes to this project will be documented in this file.
 - Fixed race condition where a stale NodeSet pod status leads to the Slurm node
   being terminated instead of drained.
 - Fixed slurm-exporter chart dependency not being in lock-step with release.
+- Fixed passing additional arguments to the slurmrestd container.
 
 ### Changed